backend/requirements.txt
# backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
huggingface-hub==0.23.0
transformers==4.41.2
torch==2.3.0 # Or tensorflow if prefer; ensure CPU or CUDA support
Pillow==10.3.0
python-dotenv==1.0.0
pydantic==2.7.1 # FastAPI dependency, good to explicitly include
celery==5.3.6 
redis==5.0.0 # For Celery backend 
# Essential for deep learning with Hugging Face models
transformers==4.41.2
Pillow==10.3.0
accelerate==0.30.1
sentencepiece==0.2.0
protobuf==4.25.3
safetensors==0.4.3
# must match torch version
torchvision==0.18.0
torchaudio==2.3.0
# Additional libraries for audio processing TTS
soundfile==0.12.1
scipy==1.13.0
librosa==0.10.1